Move non-widget code out of widget folder
